Send payload+fin in one go
Send the payload using socket.end so we immediately close our end of the connection (we're not sending anything else afterwards). This doesn't stop us from receiving the reply (the connection is half-closed).  The device calls are much faster this way. 
Also, don't clear the timeout when we get the reply, but wait until we get the close event. This stops the connection hanging if the remote end fails to shutdown the TCP connection correctly (which seems to happen occasionally with some devices).